requirements.txt
tensorflow-cpu==2.10.0
gradio==4.0.0
numpy
pillow
